refactor(moq-video)!: replace pixel_buffer() with a total into_pixel_buffer()
The borrow accessor was the wrong shape. `into_i420()` is total (it downloads
when the frame is on the GPU), but `pixel_buffer()` returned `Option<&_>` and
gave up when the frame was on the CPU. A renderer hitting `None` then had to
write an I420 -> CVPixelBuffer upload, which the crate already contained: the
VideoToolbox encoder's own CPU fallback.

So mirror `into_i420()` instead. `into_pixel_buffer()` is a retain when the
frame is already a surface and an upload when it isn't, and never fails to
produce something drawable. The upload moves out of the encode backend into
`frame::macos::upload_i420`, next to the `download_i420` it inverts, and both
callers share it.

Format varies by path (NV12 from hardware decode, planar I420 from an upload),
which is documented: CoreVideo consumers query the format anyway.

Not shipping a `#[non_exhaustive]` surface enum: variants would be platform-cfg
gated, so a match still needs cfg arms and gains nothing over per-platform
conversions, and a conversion names its target type by definition. Revisit as a
residency query when Linux lands, where CUDA and dmabuf coexist and "which am I
holding" becomes a real question.
