+++ rs/moq-video/README.md
@@ -58,10 +58,11 @@ GPU-less builder and still starts on a GPU-less machine.
 `decode::Consumer` (the mirror of `moq_audio::decode::Consumer`) subscribes to an
 H.264, H.265, or AV1 track and returns raw frames. A hardware-decoded frame stays
 on the GPU: feeding it back to `encode::Encoder` keeps it there (the transcode
-path), while `into_i420()` downloads it. The off-by-default `surface` feature
-exposes the underlying handle so you can draw it yourself, which today means
-macOS only (`decode::Frame::pixel_buffer` hands back the `CVPixelBuffer`);
-enabling it on Windows or Linux compiles but adds nothing yet. Backends are tried
+path), while `into_i420()` downloads it. The off-by-default `surface` feature adds
+`decode::Frame::into_pixel_buffer()`, the mirror of `into_i420()`: free for a
+hardware-decoded frame (already a `CVPixelBuffer`) and an upload for a CPU one, so
+you always get something to draw. That is macOS only today; enabling the feature
+on Windows or Linux compiles but adds nothing yet. Backends are tried
 hardware-first, like encode:
 
 | Codec | Software | macOS | Windows | Linux |
