+++ rs/moq-video/src/encode/backend/videotoolbox.rs
@@ -25,11 +25,7 @@ use objc2_core_media::{
 	CMFormatDescription, CMSampleBuffer, CMTime, CMVideoFormatDescriptionGetH264ParameterSetAtIndex,
 	CMVideoFormatDescriptionGetHEVCParameterSetAtIndex, kCMTimeInvalid, kCMVideoCodecType_H264, kCMVideoCodecType_HEVC,
 };
-use objc2_core_video::{
-	CVImageBuffer, CVPixelBuffer, CVPixelBufferCreate, CVPixelBufferGetBaseAddressOfPlane,
-	CVPixelBufferGetBytesPerRowOfPlane, CVPixelBufferLockBaseAddress, CVPixelBufferLockFlags,
-	CVPixelBufferUnlockBaseAddress, kCVPixelFormatType_420YpCbCr8Planar,
-};
+use objc2_core_video::CVImageBuffer;
 use objc2_video_toolbox::{
 	VTCompressionSession, VTEncodeInfoFlags, VTSessionSetProperty, kVTCompressionPropertyKey_AllowFrameReordering,
 	kVTCompressionPropertyKey_AverageBitRate, kVTCompressionPropertyKey_ExpectedFrameRate,
@@ -41,7 +37,7 @@ use objc2_video_toolbox::{
 use super::super::encoder::{Codec, Config};
 use super::Backend;
 use crate::Error;
-use crate::frame::{Frame, I420};
+use crate::frame::Frame;
 
 pub(crate) const NAME: &str = "videotoolbox";
 
@@ -162,7 +158,7 @@ impl Backend for VideoToolbox {
 		// Zero-copy when the capture handed us a surface; otherwise upload I420.
 		let pixel_buffer = match frame {
 			Frame::Surface(surface) => surface.buffer.clone(),
-			Frame::I420(i420) => make_pixel_buffer(i420)?,
+			Frame::I420(i420) => crate::frame::macos::upload_i420(i420)?,
 		};
 		let image: &CVImageBuffer = &pixel_buffer;
 
@@ -388,57 +384,6 @@ fn split_avcc(mut data: &[u8], length_size: usize) -> Vec<&[u8]> {
 		data = rest;
 	}
 	out
-}
-
-/// Allocate a planar I420 `CVPixelBuffer` and copy the frame into it (the CPU
-/// fallback, when capture didn't hand us a surface).
-fn make_pixel_buffer(frame: &I420) -> Result<CFRetained<CVPixelBuffer>, Error> {
-	let (w, h) = (frame.width as usize, frame.height as usize);
-	let (cw, ch) = (w / 2, h / 2);
-
-	let mut ptr: *mut CVPixelBuffer = ptr::null_mut();
-	let status = unsafe {
-		CVPixelBufferCreate(
-			None,
-			w,
-			h,
-			kCVPixelFormatType_420YpCbCr8Planar,
-			None,
-			NonNull::new(&mut ptr).unwrap(),
-		)
-	};
-	let buffer = NonNull::new(ptr)
-		.filter(|_| status == 0)
-		.map(|p| unsafe { CFRetained::from_raw(p) })
-		.ok_or_else(|| Error::Codec(anyhow::anyhow!("CVPixelBufferCreate failed: {status}")))?;
-
-	let flags = CVPixelBufferLockFlags(0);
-	let status = unsafe { CVPixelBufferLockBaseAddress(&buffer, flags) };
-	if status != 0 {
-		return Err(Error::Codec(anyhow::anyhow!(
-			"CVPixelBufferLockBaseAddress failed: {status}"
-		)));
-	}
-
-	copy_plane(&buffer, 0, frame.y(), w, h);
-	copy_plane(&buffer, 1, frame.u(), cw, ch);
-	copy_plane(&buffer, 2, frame.v(), cw, ch);
-
-	unsafe { CVPixelBufferUnlockBaseAddress(&buffer, flags) };
-	Ok(buffer)
-}
-
-/// Copy a tightly-packed source plane into a pixel-buffer plane, honoring its
-/// (possibly padded) row stride.
-fn copy_plane(buffer: &CVPixelBuffer, plane: usize, src: &[u8], row_bytes: usize, rows: usize) {
-	let base = CVPixelBufferGetBaseAddressOfPlane(buffer, plane) as *mut u8;
-	let stride = CVPixelBufferGetBytesPerRowOfPlane(buffer, plane);
-	for y in 0..rows {
-		unsafe {
-			let dst = base.add(y * stride);
-			ptr::copy_nonoverlapping(src[y * row_bytes..].as_ptr(), dst, row_bytes);
-		}
-	}
 }
 
 fn force_keyframe_dict() -> Result<CFRetained<CFDictionary>, Error> {
